Dialog de opcoes para algoritmo Penso que sem bugs... Pus os edge que fazem parte duma solucao mais espessos para se identificar melhor

+++ Ta todo la dentro/src/dataStructure/city/infraStructure/Vehicle.java
@@ -53,7 +53,7 @@ public class Vehicle {
 		return maxSpeed;
 	}
 
-	public void setCurrentFuel(int fuel) {
+	public void setCurrentFuel(double fuel) {
 		if(fuel > maxCapacity){
 			currentFuel = maxCapacity;
 		} else {
